First chunk of cleaning up PC9801: move state to includes, make variables to have sane accessors (nw)
