+++ src/devices/bus/cbus/pc9801_cbus.cpp
@@ -3,7 +3,7 @@
 /**********************************************************************
 
     CCBUS Slot interface for PC-98xx family
-
+	
 **********************************************************************/
 
 #include "emu.h"
